fix: Relocate enumeration classes to comply with CODING_RULE_STYLE_00008 and CODING_RULE_STYLE_00009
- Moved FlexrayChannelName from CoreTopology.py to FlexrayTopology.py
- Moved CommunicationDirectionType from CoreTopology.py to CoreCommunication/__init__.py
- Moved IPduSignalProcessingEnum from CoreTopology.py to CoreCommunication/__init__.py
- Updated imports in DataMapping.py, CoreTopology.py, and test files
- Removed empty FlexrayTopology/ directory (CODING_RULE_STYLE_00008 violation)
- Used TYPE_CHECKING and string type annotations to avoid circular imports

Class mapping progress:
- Before: 734 correct (37.9%), 50 wrong
- After: 737 correct (38.0%), 47 wrong
- Progress: 7 of 54 violations fixed (13.0%)
