+++ src/armodel/models/M2/AUTOSARTemplates/SystemTemplate/Fibex/Fibex4Flexray/FlexrayTopology.py
@@ -1,8 +1,25 @@
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import Boolean, Float, Integer, PositiveInteger
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import PositiveUnlimitedInteger, TimeValue
+from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import AREnum
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ArObject import ARObject
 from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.FibexCore.CoreTopology import CommunicationCluster, CommunicationConnector
 from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.FibexCore.CoreTopology import CommunicationController
+
+
+class FlexrayChannelName(AREnum):
+    """
+    Enumeration defining names for FlexRay channels,
+    specifying the available channel designations
+    in FlexRay communication systems.
+    """
+    CHANNEL_A = "channelA"
+    channel_B = "channelB"
+
+    def __init__(self):
+        super().__init__([
+            FlexrayChannelName.CHANNEL_A,
+            FlexrayChannelName.channel_B
+        ])
 
 
 class FlexrayCommunicationController(CommunicationController):
